Provide pipeline information
While the majority of the information about the pipeline, individual pipeline steps, how programs are called etc. properly lives in the zika-seq repo, I wanted there to be a brief description of how we derived consensus genomes available in the README here, along with a link out to the zika-seq repo, to provide a bit of clarity about how we get genomes.