Add auspice JSONs
One of the cool features now of nextstrain auspice is that, if you have Auspice formatted JSON files and you put them in an auspice directory in a git repo, you can automatically access a nextstrain visualization of that build via web address. Importantly, this ensures that there is a nextstrain visualization compatible with the dataset used in your a manuscript's analysis. Having this kind of static nextstrain build is especially important in this case as we are using the nextstrain pipeline as our primary analytic and visualization pipeline.