Grab latest genomes from Fauna
Latest zika download from fauna. This pull grabs _all_ publicly available genomes off of ViPR, Genbank, and specified git repos. Notably while these are all genomes in the public domain, we still need to seek permission to use some of them in a published analysis if that genome has never been published on. As such, this file is not the input data file for all subsequent analyses, but rather the first pull down. A subset of the genomes in this file will be the dataset for our analysis.