Add file with permission descriptions
This file contains all the manually curated permission information. If a genome has already been published on, that is noted and the pubmed ID of the paper is also given. If a genome wasn't published on, but we received written confirmation that it would be okay to include that genome in the analysis, then that is also noted. Some genomes are not published on, and also have no approval to use them. These genomes without permission will not be used in our analysis, even though they are in the public domain.