update auspice files
Previously the JSONs here were the same as for the build of the live nextstrain build. However, given that the zika-colombia analyses will be published, the augur analysis needed to be done on a subset of genomes that can be included in published analyses. This commit pushes Auspice JSONs up to the repo that were made from an augur build using only genomes that can be published on.