remove sequences that deviate from clock
When I did the nextstrain auspice visualization, I found three samples that were very deviant from the molecular clock estimated from the rest of the dataset. Given that this kind of deviation can be a sign of contamination or incorrect metadata, I have removed these three sequences from the analysis.