update env file
Basemap, the python program I'm using to plot sampling locations with, is a little finicky and some important functionality is broken in different releases of the code. This env file specifies the exact version and channel to source Basemap from, and also contains the code to ensure that Jupyter notebook will run this specific environment.