Update fasta headers
I've added more information to the fasta headers so that it is clear what MinION chemistry was used for sequencing, how many genomes were multiplexed on the run, and to indicate porechop stringency. These sequences also have site 10375 masked with an N, as nanopolish seems to consistently fail to call an appropriate SNP at this site.