N-mask site 10375
Nanopolish seems to have some issue where at this site specifically it calls the reference rather than a SNP (despite strong evidence in the BAM stack for a SNP, and the fact that all other sequences in the Americas differ from French Polynesian sequences at this site). Since this seems to artifactual, I've N-masked this site in our Colombian sequences.