Note docker memory alloc in README
I was finding that nextstrain cli would stochastically fail because mafft would crash. The issue is outlined in greater detail in nextstrain/cli issue 20 (https://github.com/nextstrain/cli/issues/20). This error could be corrected by increasing the memory allocation to Docker. This error could likely crop up for anyone else running on a Mac (Docker limits memory allocation for Mac but not Linux machines), so I wanted it addressed in the README.