add pipeline run files
Adding all of the results files that get generated while running the pipeline to the repo. This includes alignments and trees and the like. Presence/absence of these files will tell Snakemake where to begin if starting from the middle of a run, so if the repo is fully cloned and the pipeline is run, then the auspice JSONs should simply get recreated (if they are not already present in the auspice directory).