update gitignore, don't track manuscript changes at this time
There's are lots of edits to the manuscript and figures going on right now so I just want this to not be tracked right now or else there's too much fuss in the staging area. I'll add manuscript files back up to GitHub once we are submission ready, and again at acceptance.