clarify Ecuadorian strain names
The Ecuadorian strain names did not include the word Ecuador, nor any standard abbreviation for the country, and therefore tip labels in figures weren't clear. I will not change this in Fauna, as the less clear strain name is the strain name chosen by the GenBank submitting authors, but I do want this clearer in my own figures.