Add supplemental analysis
We wanted to explore what the tree topology looks like if you include genomes with less than 50% coverage. The creation of this directory (along with it's own files and Snakefile) allows the exact same analysis to be run, albeit on the slightly different input fasta that contains the lower quality genomes. Once moved into the supplemental-analysis directory, building the analysis works exactly the same as it does in the top-level directory.