Describe how to get bezier curve-free figs
I think this information is important to document (for other who want to do something similar, and for myself so I know what I did in the future). It probably belongs in a different README, like one a figures directory, but I'm putting it here now to get it committed.