Add header info to README
In the ZIKA-COL-* fasta files, the fasta headers contain all of the epi metadata as well as information about the bioinformatic processing that was performed on the sample. However, it might not have been totally clear what all of these fields were. This commit adds this information into the README in order to make it clear what the different header fields represent.