Add functionality to make some different types of input fastas
We were having issues with the MAFFT alignment doing weird stuff with sequences that had lots and lots of N's. To troubleshoot how to get around this, I needed to make a couple of different input fasta files, one where the N's are swapped for gaps, and one fasta where any non-Colombia sequences that was more than 20% N was removed from the dataset.