Add date parsing from JSON files
I want to plot sampling dates for Colombian genomes on the epi curve to show how genomic sampling intensity relates to incidence. Previously I had just looked at dates on nextstrain and hardcoded them in, but that's not at all reproducible. This addition of code will parse those dates in now from JSON files that are made as the analysis is built.