update after re-build
The purpose of this rebuild was to fix an error that used to occur in TreeTime, that made some branches with minimal genetic changes be much too long in the temporally resolved version of the tree. The new tree uses the new treetime code whereby temporally resolved branch lengths do sync with genetic distance branch lengths.