add division level color-by feature
To do a bit more analysis at the within-colombia level, I wanted to add a color-by feature to the nextstrain build that would allow us to look at which department within Colombia a genome was sampled from. This required adding this information to the input data, altering the auspice_config to add a color by division flag, and adding the colors that I want to the colors.tsv file.