Update snakefile to fully working version
Previously I had stubbed out a Snakefile with just the barebones of what I needed to think about in terms of expansion through all the sequence and trial combinations, as well as some file concatenation I needed to do. I have since gotten the Snakefile to fully specify the workflow as it is running on AWS batch, and that is what is being committed here.