Fix subsampling order bug
I realized that I had nested my 3 for loops incorrectly, and thus that the same subsampled sequences were getting pulled for each trial. I re-ordered the for loops such that the subsampling function is called on the most nested loop, which fixes this issue (and I checked this by looking at the file contents).